Add event to logic folder
